added some journal endpoints, code refactoring
